Use decorator pattern.
